added new logic to modularize adding font loader and font to scene
