+++ threejs-ng-snake/src/app/canvas-comp/canvas-comp.component.ts
@@ -6,6 +6,7 @@ import { WordApiService } from '../word-api.service';
 // import { FormBuilder } from '@angular/forms'
 import { ObjBuilderService } from '../services/obj-builder.service'
 import { SceneHelperService } from '../services/scene-helper.service'
+import { FontBuilderService } from '../services/font-builder.service'
 
 @Component({
     selector: 'app-canvas-comp',
@@ -23,16 +24,16 @@ export class CanvasCompComponent implements OnInit {
     public camera: THREE.PerspectiveCamera;
     public renderer: any;
     public start: any;
-//     public controls: any;
     public controls: any;
-    public loader: FontLoader;
+//     public loader: FontLoader;
     public wordGet: any;
     public axesHelper: THREE.AxesHelper;
     public gridHelper: THREE.GridHelper;
 
     constructor(private wordService: WordApiService,
                 private builderService: ObjBuilderService,
-                private sceneService: SceneHelperService
+                private sceneService: SceneHelperService,
+                private fontService: FontBuilderService
                 ) {
         this.scene = new THREE.Scene();
         // todo new logic axes and grid; could be modularized
@@ -48,45 +49,44 @@ export class CanvasCompComponent implements OnInit {
 
         this.camera = new THREE.PerspectiveCamera(60, 800 / 600);
         this.start = -1;
-//         this.initLights()
         this.sceneService.initLights(this.scene)
-//         this.initFog()
         this.sceneService.initFog(this.scene)
         //for font
-        this.loader = new FontLoader();
-        this.addFont("Hello\nWorld")
+//         this.loader = new FontLoader();
+//         this.addFont("Hello\nWorld")
+        this.fontService.addFont("Hello\nWorld", this.scene)
         // necessary to enable "this" keyword to work correctly inside animate
         this.animate = this.animate.bind(this);
     }
 
 
 
-    addFont(msg: string) : void {
-        // text
-        // https://threejs.org/examples/?q=text#webgl_geometry_text_shapes
-        // https://github.com/mrdoob/three.js/blob/master/examples/webgl_geometry_text_shapes.html
-        const fontUri = '..\\assets\\helvetiker_regular.typeface.json'
-        this.loader.load(fontUri, font => {
-            const fontColor = new THREE.Color('rgb(0, 255, 0)');
-            const matLite = new THREE.MeshBasicMaterial({
-                color: fontColor,
-                transparent: true,
-                opacity: .5,
-                side: THREE.DoubleSide
-            });
-            const message = msg
-            const shape = font.generateShapes(message, .2);
-            const textGeo = new THREE.ShapeGeometry(shape);
-            textGeo.computeBoundingBox();
-            // do some logic for move center of text using bounding box
-            const text = new THREE.Mesh(textGeo, matLite);
-            text.name = 'wordName';
-            text.position.z = 1;
-            text.position.y = .5;
-            text.position.x = -.5;
-            this.scene.add(text);
-        });
-    }
+//     addFont(msg: string) : void {
+//         // text
+//         // https://threejs.org/examples/?q=text#webgl_geometry_text_shapes
+//         // https://github.com/mrdoob/three.js/blob/master/examples/webgl_geometry_text_shapes.html
+//         const fontUri = '..\\assets\\helvetiker_regular.typeface.json'
+//         this.loader.load(fontUri, font => {
+//             const fontColor = new THREE.Color('rgb(0, 255, 0)');
+//             const matLite = new THREE.MeshBasicMaterial({
+//                 color: fontColor,
+//                 transparent: true,
+//                 opacity: .5,
+//                 side: THREE.DoubleSide
+//             });
+//             const message = msg
+//             const shape = font.generateShapes(message, .2);
+//             const textGeo = new THREE.ShapeGeometry(shape);
+//             textGeo.computeBoundingBox();
+//             // do some logic for move center of text using bounding box
+//             const text = new THREE.Mesh(textGeo, matLite);
+//             text.name = 'wordName';
+//             text.position.z = 1;
+//             text.position.y = .5;
+//             text.position.x = -.5;
+//             this.scene.add(text);
+//         });
+//     }
 
 
     // @ts-ignore
@@ -98,7 +98,6 @@ export class CanvasCompComponent implements OnInit {
             this.start = timestamp;
         }
         const elapsed = timestamp - this.start;
-//         const testCube = this.scene.getObjectByName('test_box');
         //     https://dustinpfister.github.io/2021/05/12/threejs-object3d-get-by-name/
         const textObj = this.scene.getObjectByName('wordName');
         /*note todo here: trying to set word based on API response; probably need to create new shape if can't find attribue to change
@@ -109,7 +108,7 @@ export class CanvasCompComponent implements OnInit {
             // todo this shouldn't be a global probably
             if(this.wordGet!=undefined){
                 this.scene.remove(textObj)
-                this.addFont(this.wordGet)
+                this.fontService.addFont(this.wordGet, this.scene)
             }
 
         }
@@ -154,7 +153,6 @@ export class CanvasCompComponent implements OnInit {
 
 
     ngOnInit(): void {
-//         this.scene.add(this.mesh);
         this.renderer = new THREE.WebGLRenderer({
             antialias: true,
             logarithmicDepthBuffer: true,
@@ -164,12 +162,10 @@ export class CanvasCompComponent implements OnInit {
         this.renderer.shadowMap.enabled = true
     // @ts-ignore
         this.renderer.setClearColor(this.scene.fog.color)
-//         this.init_cameras();
         this.sceneService.initCameras(this.scene, this.camera)
         this.controls = this.sceneService.initControls(this.scene, this.camera)
         this.window_set_size();
         this.window_size_listener();
-//         this.initBoxes();
         this.builderService.initBoxes(this.shapesArray, this.scene)
 
         requestAnimationFrame(this.animate);
