+++ threejs-ng-snake/src/app/services/font-builder.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FontBuilderService } from './font-builder.service';
+
+describe('FontBuilderService', () => {
+  let service: FontBuilderService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(FontBuilderService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
